small refactor of resoure isolation group utilities
